+++ packages/apps-config/src/api/spec/index.ts
@@ -7,6 +7,7 @@ import canvas from './canvas';
 import centrifugeChain from './centrifuge-chain';
 import crab from './crab';
 import crust from './crust';
+import crustParachain from './crust-parachain';
 import testPara from './cumulus-test-parachain';
 import darwinia from './darwinia';
 import darwiniaParachain from './darwinia-parachain';
@@ -21,6 +22,7 @@ import jupiter from './jupiter';
 import kilt from './kilt';
 import nodle from './nodle';
 import phala from './phala';
+import phalaParachain from './phala-parachain';
 import plasm from './plasm';
 import plasmParachain from './plasm-parachain';
 import polkadex from './polkadex';
@@ -41,6 +43,7 @@ export default {
   canvas,
   'centrifuge-chain': centrifugeChain,
   crust,
+  'crust-parachain': crustParachain,
   'cumulus-test-parachain': testPara,
   'darwinia-parachain': darwiniaParachain,
   'dock-main-runtime': dock,
@@ -54,6 +57,7 @@ export default {
   'mashnet-node': kilt,
   'node-polkadex': polkadex,
   'nodle-chain': nodle,
+  'phala-collator': phalaParachain,
   'phala-node': phala,
   plasm,
   'plasm-parachain': plasmParachain,