+++ packages/apps-config/src/api/specVersioned/polkabtc.ts
@@ -5,4 +5,8 @@ import type { OverrideBundleDefinition } from '@polkadot/types/types';
 
 import definitions from '@interlay/polkabtc-types';
 
-export default definitions as OverrideBundleDefinition;
+// This is a temp work-around to add the interfaces key to the actual types
+// With the next release of the API this should be able tp be added to the base
+const extended = definitions as OverrideBundleDefinition;
+
+export default extended;