+++ packages/page-accounts/src/modals/Transfer.tsx
@@ -8,7 +8,8 @@ import BN from 'bn.js';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
-import { InputAddress, InputBalance, MarkWarning, Modal, Toggle, TxButton } from '@polkadot/react-components';
+import { checkAddress } from '@polkadot/phishing';
+import { InputAddress, InputBalance, MarkError, MarkWarning, Modal, Toggle, TxButton } from '@polkadot/react-components';
 import { useApi, useCall } from '@polkadot/react-hooks';
 import { Available } from '@polkadot/react-query';
 import { BN_ZERO, isFunction } from '@polkadot/util';
@@ -26,6 +27,19 @@ function isRefcount (accountInfo: AccountInfoWithProviders | AccountInfoWithRefC
   return !!(accountInfo as AccountInfoWithRefCount).refcount;
 }
 
+async function checkPhishing (_senderId: string | null, recipientId: string | null): Promise<[string | null, string | null]> {
+  return [
+    // not being checked atm
+    // senderId
+    //   ? await checkAddress(senderId)
+    //   : null,
+    null,
+    recipientId
+      ? await checkAddress(recipientId)
+      : null
+  ];
+}
+
 function Transfer ({ className = '', onClose, recipientId: propRecipientId, senderId: propSenderId }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const { api } = useApi();
@@ -36,6 +50,7 @@ function Transfer ({ className = '', onClose, recipientId: propRecipientId, send
   const [[maxTransfer, noFees], setMaxTransfer] = useState<[BN | null, boolean]>([null, false]);
   const [recipientId, setRecipientId] = useState<string | null>(propRecipientId || null);
   const [senderId, setSenderId] = useState<string | null>(propSenderId || null);
+  const [[, recipientPhish], setPhishing] = useState<[string | null, string | null]>([null, null]);
   const balances = useCall<DeriveBalancesAll>(api.derive.balances.all, [senderId]);
   const accountInfo = useCall<AccountInfoWithProviders | AccountInfoWithRefCount>(api.query.system.account, [senderId]);
 
@@ -64,6 +79,12 @@ function Transfer ({ className = '', onClose, recipientId: propRecipientId, send
       setMaxTransfer([null, false]);
     }
   }, [api, balances, recipientId, senderId]);
+
+  useEffect((): void => {
+    checkPhishing(senderId, recipientId)
+      .then(setPhishing)
+      .catch(console.error);
+  }, [recipientId, senderId]);
 
   const noReference = accountInfo
     ? isRefcount(accountInfo)
@@ -117,6 +138,9 @@ function Transfer ({ className = '', onClose, recipientId: propRecipientId, send
                 onChange={setRecipientId}
                 type='allPlus'
               />
+              {recipientPhish && (
+                <MarkError content={t<string>('The recipient is associated with a known phishing site on {{url}}', { replace: { url: recipientPhish } })} />
+              )}
             </Modal.Column>
             <Modal.Column>
               <p>{t<string>('The beneficiary will have access to the transferred fees when the transaction is included in a block.')}</p>
@@ -192,7 +216,7 @@ function Transfer ({ className = '', onClose, recipientId: propRecipientId, send
         <TxButton
           accountId={senderId}
           icon='paper-plane'
-          isDisabled={!hasAvailable || !recipientId || !amount}
+          isDisabled={!hasAvailable || !recipientId || !amount || !!recipientPhish}
           label={t<string>('Make Transfer')}
           onStart={onClose}
           params={