+++ packages/page-bounties/src/BountyActions/index.tsx
@@ -6,12 +6,14 @@ import type { Balance, BlockNumber, BountyIndex, BountyStatus } from '@polkadot/
 
 import React, { useCallback, useMemo } from 'react';
 
+import { getBountyStatus } from '../helpers';
 import BountyClaimAction from './BountyClaimAction';
 import BountyCuratorProposedActions from './BountyCuratorProposedActions';
 import BountyInitiateVoting from './BountyInitiateVoting';
+import CloseBounty from './CloseBounty';
 import ExtendBountyExpiryAction from './ExtendBountyExpiryAction';
-import { getBountyStatus } from './helpers';
 import ProposeCuratorAction from './ProposeCuratorAction';
+import SlashCurator from './SlashCurator';
 
 interface Props {
   bestNumber: BlockNumber;
@@ -25,8 +27,9 @@ interface Props {
 export function BountyActions ({ bestNumber, description, index, proposals, status, value }: Props): JSX.Element {
   const updateStatus = useCallback(() => getBountyStatus(status), [status]);
 
-  const { beneficiary, curator, unlockAt } = updateStatus();
+  const { beneficiary, curator, unlockAt, updateDue } = updateStatus();
 
+  const blocksUntilUpdate = useMemo(() => updateDue?.sub(bestNumber), [bestNumber, updateDue]);
   const blocksUntilPayout = useMemo(() => unlockAt?.sub(bestNumber), [bestNumber, unlockAt]);
 
   return (
@@ -35,20 +38,22 @@ export function BountyActions ({ bestNumber, description, index, proposals, stat
         <BountyInitiateVoting
           index={index}
           proposals={proposals}
-        />}
-      {status.isFunded && (
+        />
+      }
+      {status.isFunded &&
         <ProposeCuratorAction
           description={description}
           index={index}
           proposals={proposals}
           value={value}
         />
-      )}
+      }
       {status.isCuratorProposed && curator &&
         <BountyCuratorProposedActions
           curatorId={curator}
           index={index}
-        />}
+        />
+      }
       {status.isActive && curator &&
           <ExtendBountyExpiryAction
             curatorId={curator}
@@ -56,12 +61,29 @@ export function BountyActions ({ bestNumber, description, index, proposals, stat
             index={index}
           />
       }
+      {(status.isFunded || status.isActive || status.isCuratorProposed) &&
+        <CloseBounty
+          index={index}
+          proposals={proposals}
+        />
+      }
       {status.isPendingPayout && beneficiary && blocksUntilPayout &&
         <BountyClaimAction
           beneficiaryId={beneficiary}
           index={index}
           payoutDue={blocksUntilPayout}
-        />}
+        />
+      }
+      {(status.isCuratorProposed || status.isActive || status.isPendingPayout) && curator && (
+        <SlashCurator
+          blocksUntilUpdate={blocksUntilUpdate}
+          curatorId={curator}
+          description={description}
+          index={index}
+          proposals={proposals}
+          status={status}
+        />
+      )}
     </>
   );
 }