+++ packages/page-bounties/src/hooks/useBounties.tsx
@@ -10,18 +10,21 @@ import { useApi, useCall } from '@polkadot/react-hooks';
 import { BalanceOf, BlockNumber, BountyIndex } from '@polkadot/types/interfaces';
 
 export type BountyApi = {
+  acceptCurator: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   approveBounty: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   bestNumber?: BlockNumber,
   bounties?: DeriveBounties,
   bountyDepositBase: BN,
   bountyIndex?: BN,
   bountyValueMinimum: BN,
+  claimBounty: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   closeBounty: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   dataDepositPerByte: BN,
   extendBountyExpiry: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   maximumReasonLength: number,
   proposeBounty: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
   proposeCurator: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
+  unassignCurator: ((...args: any[]) => SubmittableExtrinsic<'promise'>);
 };
 
 export function useBounties (): BountyApi {
@@ -36,22 +39,28 @@ export function useBounties (): BountyApi {
   const dataDepositPerByte = (constsBase.dataDepositPerByte as BalanceOf).toBn();
   const proposeBounty = (api.tx.bounties || api.tx.treasury).proposeBounty;
   const proposeCurator = (api.tx.bounties || api.tx.treasury).proposeCurator;
+  const claimBounty = (api.tx.bounties || api.tx.treasury).claimBounty;
+  const acceptCurator = (api.tx.bounties || api.tx.treasury).acceptCurator;
   const approveBounty = (api.tx.bounties || api.tx.treasury).approveBounty;
   const closeBounty = (api.tx.bounties || api.tx.treasury).closeBounty;
   const extendBountyExpiry = (api.tx.bounties || api.tx.treasury).extendBountyExpiry;
+  const unassignCurator = (api.tx.bounties || api.tx.treasury).unassignCurator;
 
   return {
+    acceptCurator,
     approveBounty,
     bestNumber,
     bounties,
     bountyDepositBase,
     bountyIndex,
     bountyValueMinimum,
+    claimBounty,
     closeBounty,
     dataDepositPerByte,
     extendBountyExpiry,
     maximumReasonLength,
     proposeBounty,
-    proposeCurator
+    proposeCurator,
+    unassignCurator
   };
 }