+++ packages/page-bounties/src/index.tsx
@@ -2,8 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useRef } from 'react';
+import styled from 'styled-components';
 
 import { Tabs } from '@polkadot/react-components';
+import { ThemeProps } from '@polkadot/react-components/types';
 
 import Bounties from './Bounties';
 import { useTranslation } from './translate';
@@ -12,9 +14,10 @@ export { default as useCounter } from './useCounter';
 
 interface Props {
   basePath: string;
+  className?: string;
 }
 
-function BountiesApp ({ basePath }: Props): React.ReactElement<Props> {
+function BountiesApp ({ basePath, className = '' }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
 
   const itemsRef = useRef([
@@ -26,7 +29,7 @@ function BountiesApp ({ basePath }: Props): React.ReactElement<Props> {
   ]);
 
   return (
-    <main className='bounties--App'>
+    <main className={`bounties--App ${className}`}>
       <header>
         <Tabs
           basePath={basePath}
@@ -38,4 +41,77 @@ function BountiesApp ({ basePath }: Props): React.ReactElement<Props> {
   );
 }
 
-export default React.memo(BountiesApp);
+export default React.memo(styled(BountiesApp)(({ theme }: ThemeProps) => `
+  && thead th {
+    padding: 0.75rem 1rem 0.57rem;
+    font-weight: bold;
+    font-size: 0.7rem;
+    line-height: 0.85rem;
+    text-transform: uppercase;
+    color: ${theme.colorLabel};
+    border: none;
+  }
+
+  thead tr {
+    border: none;
+    background: none;
+  }
+
+  tbody td {
+    padding: 1.4rem 1rem;
+    vertical-align: baseline;
+    font-size: 1rem;
+    line-height: 1.7rem;
+    color: ${theme.theme === 'dark' ? '#eeeeee' : '#1a1b20'}
+    border-bottom: 1px solid ${theme.theme === 'dark' ? '#2f313c' : '#dfdfdf'};
+
+    &:first-child {
+      border-left: 1px solid  ${theme.theme === 'dark' ? '#2f313c' : '#dfdfdf'};
+    }
+
+    &:last-child {
+      border-right: 1px solid  ${theme.theme === 'dark' ? '#2f313c' : '#dfdfdf'};
+    }
+  }
+
+  tbody tr {
+    &:nth-child(odd) {
+      background: ${theme.theme === 'dark' ? '#111218' : '#fff'};
+    }
+
+    &:nth-child(odd) td {
+      vertical-align: middle;
+    }
+
+    &:nth-child(even) {
+      background: ${theme.bgPage};
+    }
+
+    &:first-child {
+      td {
+        border-top: 1px solid  ${theme.theme === 'dark' ? '#2f313c' : '#dfdfdf'};
+      }
+
+      td:first-child {
+        border-top-left-radius: 0.285rem;
+      }
+
+      td:last-child {
+        border-top-right-radius: 0.285rem;
+      }
+    }
+
+    &:last-child {
+      td {
+        border-bottom: 1px solid  ${theme.theme === 'dark' ? '#2f313c' : '#dfdfdf'};
+      }
+
+      td:first-child {
+        border-bottom-left-radius: 0.285rem;
+      }
+      td:last-child {
+        border-bottom-right-radius: 0.285rem;
+      }
+    }
+  }
+`));