+++ packages/page-treasury/src/Overview/Summary.tsx
@@ -7,8 +7,7 @@ import BN from 'bn.js';
 import React from 'react';
 
 import { CardSummary, SummaryBox } from '@polkadot/react-components';
-import { useApi, useCall } from '@polkadot/react-hooks';
-import { useTreasury } from '@polkadot/react-hooks/useTreasury';
+import { useApi, useCall, useTreasury } from '@polkadot/react-hooks';
 import { FormatBalance } from '@polkadot/react-query';
 import { formatNumber } from '@polkadot/util';