+++ packages/react-components/src/Table/Head.tsx
@@ -15,9 +15,10 @@ interface Props {
   filter?: React.ReactNode;
   header?: (null | undefined | HeaderDef)[];
   isEmpty: boolean;
+  hasTitle: boolean;
 }
 
-function Head ({ className = '', filter, header, isEmpty }: Props): React.ReactElement<Props> | null {
+function Head ({ className = '', filter, hasTitle, header, isEmpty }: Props): React.ReactElement<Props> | null {
   if (!header?.length) {
     return null;
   }
@@ -37,7 +38,7 @@ function Head ({ className = '', filter, header, isEmpty }: Props): React.ReactE
             key={index}
             onClick={onClick}
           >
-            {index === 0
+            {index === 0 && hasTitle
               ? (
                 <h1>
                   <Icon