+++ packages/react-components/src/Table/index.tsx
@@ -16,6 +16,7 @@ interface TableProps {
   filter?: React.ReactNode;
   footer?: React.ReactNode;
   header?: [React.ReactNode?, string?, number?, (() => void)?][];
+  hasTitle: boolean;
   isFixed?: boolean;
   legend?: React.ReactNode;
 }
@@ -31,7 +32,7 @@ function extractKids (children: React.ReactNode): [boolean, React.ReactNode] {
   return [isEmpty, isEmpty ? null : kids];
 }
 
-function Table ({ children, className = '', empty, emptySpinner, filter, footer, header, isFixed, legend }: TableProps): React.ReactElement<TableProps> {
+function Table ({ children, className = '', empty, emptySpinner, filter, footer, hasTitle = true, header, isFixed, legend }: TableProps): React.ReactElement<TableProps> {
   const [isEmpty, kids] = extractKids(children);
 
   return (
@@ -40,6 +41,7 @@ function Table ({ children, className = '', empty, emptySpinner, filter, footer,
       <table className={`${(isFixed && !isEmpty) ? 'isFixed' : 'isNotFixed'} highlight--bg-faint`}>
         <Head
           filter={filter}
+          hasTitle={hasTitle}
           header={header}
           isEmpty={isEmpty}
         />